add small script
